requirements.txt
aiml==0.9.2
alembic==1.12.1
annotated-types==0.5.0
anyio==3.7.1
apyori==1.1.2
click==8.1.3
colorama==0.4.6
cvzone==1.5.6
cycler==0.11.0
exceptiongroup==1.3.0
fastapi==0.103.2
fonttools==4.33.3
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib-metadata==6.1.0
importlib-resources==5.12.0
joblib==1.1.0
kiwisolver==1.4.2
Mako==1.2.4
MarkupSafe==2.1.5
matplotlib==3.5.1
nltk==3.8.1
numpy==1.21.6
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
Pillow==9.1.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.1
regex==2022.10.31
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
sklearn==0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.27.0
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.7.1
uvicorn==0.22.0
zipp==3.15.0
